requirements.txt
altair==5.2.0
annotated-types==0.7.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
azure-functions==1.20.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cryptography==42.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.9.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.13.1
Flask==2.3.2
fonttools==4.40.0
fqdn==1.5.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.0
jsonpointer==2.4
jsonschema==4.17.3
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mpmath==1.3.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.2.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.25.0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20221105
pdfplumber==0.10.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.5
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.8.0
pydantic_core==2.20.0
pydeck==0.8.1b0
Pygments==2.15.1
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.0
pypdfium2==4.26.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2022.7.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.32.1
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.11.1
Send2Trash==1.8.2
sentence-transformers==2.3.1
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
streamlit==1.30.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.15.1
toml==0.10.2
toolz==0.12.0
torch==2.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.37.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
uri-template==1.3.0
urllib3==2.0.6
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.7
zipp==3.17.0
